+++ lbz/resource.py
@@ -1,11 +1,11 @@
 #!/usr/local/bin/python3.8
 # coding=utf-8
-"""
-Resource Handler.
-"""
+"""Resource Handler."""
 import traceback
 from os import environ as env
 from typing import Union
+
+from multidict import CIMultiDict
 
 from lbz.authentication import User
 from lbz.authz import Authorizer
@@ -29,13 +29,14 @@ class Resource:
 
     def __init__(self, event: dict):
         self._load_configuration()
+        self.urn = event["path"]  # TODO: Variables should match corresponding event fields
         self.path = event.get("requestContext", {}).get("resourcePath")
-        self.uids = event.get("pathParameters") if event.get("pathParameters") is not None else {}
+        self.path_params = event.get("pathParameters", {})
         self.method = event["requestContext"]["httpMethod"]
-        headers = event["headers"]
+        headers = CIMultiDict(event.get("headers", {}))
         self.request = Request(
             headers=headers,
-            uri_params=self.uids,
+            uri_params=self.path_params,
             method=self.method,
             body=event["body"],
             context=event["requestContext"],
@@ -43,7 +44,7 @@ class Resource:
             is_base64_encoded=event.get("isBase64Encoded", False),
             query_params=event["multiValueQueryStringParameters"],
         )
-        if authorization := headers.get("Authorization", headers.get("authorization")):
+        if authorization := headers.get("Authorization"):
             self._authorizer.set_policy(authorization)
         else:
             self._authorizer.reset_policy()
@@ -51,12 +52,12 @@ class Resource:
     def __call__(self) -> dict:
         try:
             if self.path is None or self.path not in self._router:
-                logger.error("Couldn't find PATH %s in current paths %s", self.path, self._router)
-                raise WrongURI()
+                logger.error("Couldn't find %s in current paths: %s", self.path, self._router)
+                raise WrongURI
             if self.method not in self._router[self.path]:
                 raise UnsupportedMethod(method=self.method)
             self.request.user = self._get_user(self.request.headers)
-            return getattr(self, self._router[self.path][self.method])(**self.uids)
+            return getattr(self, self._router[self.path][self.method])(**self.path_params)
         except LambdaFWException as e:
             logger.format_error(e)
             if self.print_traceback and 500 <= e.status_code < 600:
@@ -64,14 +65,14 @@ class Resource:
             return e.get_resp()
 
     def __repr__(self):
-        return f"<Resource {self.method} @ {self.path} UIDS: {self.uids}>"
+        return f"<Resource {self.method} @ {self.urn} >"
 
     def _load_configuration(self) -> None:
         self.print_traceback = bool(int(env.get("PRINT_TRACEBACK", "0")))
         self.auth_enabled = env.get("ALLOWED_PUBLIC_KEYS") or env.get("ALLOWED_AUDIENCES")
 
-    def _get_user(self, headers: dict) -> Union[None, User]:
-        authentication = headers.get("Authentication", headers.get("authentication"))
+    def _get_user(self, headers: CIMultiDict) -> Union[None, User]:
+        authentication = headers.get("Authentication")
         if authentication and self.auth_enabled:
             return User(authentication)
         elif authentication: